Move useSort (now useTableSort) into its own module
